Windows-proofing!

+++ satella/posix/__init__.py
@@ -3,6 +3,7 @@ POSIX things
 """
 
 import os
+import sys
 
 from .daemon import daemonize
 from .pidlock import PIDFileLock, LockIsHeld
@@ -24,7 +25,12 @@ def is_running_as_root() -> bool:
     Checks whether EUID is 0
 
     :return: bool
+    :raises OSError: called on Windows!
     """
+
+    if sys.platform.startswith('win'):
+        raise OSError('Routine unavailable on Windows!')
+
     return os.geteuid() == 0
 
 
@@ -33,7 +39,12 @@ def suicide(kill_entire_pg: bool = True) -> None:
     Kill self.
 
     :param kill_entire_pg: whether to kill entire PG if a session leader
+    :raises NotImplementedError: called on Windows!
     """
+
+    if sys.platform.startswith('win'):
+        raise NotImplementedError('suicide() not yet supported on Windows! (see issue #37)')
+
     mypid = os.getpid()
     kill = os.killpg if kill_entire_pg and os.getpgid(0) == mypid else os.kill
     kill(mypid, 9)
